Tiny optimization: If adding an empty list of objects to an SQL store, just return -- don't set up the entire connection and try to access the database.

+++ storage/src/main/java/edu/uiuc/ncsa/security/storage/sql/SQLStore.java
@@ -533,6 +533,7 @@ public abstract class SQLStore<V extends Identifiable> extends SQLDatabase imple
     }
 
     public boolean remove(List<Identifiable> objects) {
+        if(objects.size() == 0) return true; // do nothing
         String inSql = String.join(",", Collections.nCopies(objects.size(), "?"));
         inSql = "(" + inSql + ")";
         String query = "DELETE FROM " + getTable().getFQTablename() + " WHERE " + getTable().getPrimaryKeyColumnName() + " IN " + inSql;
